pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "creditlens"
version = "1.0.0"
description = "CreditLens — AI Credit Risk Underwriting Environment (OpenEnv)"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "CreditLens Team" }]

dependencies = [
    "faker>=24.0.0",
    "sdv>=1.12.0",
    "duckdb>=0.10.0",
    "pyarrow>=15.0.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.4.0",
    "xgboost>=2.0.0",
    "shap>=0.45.0",
    "networkx>=3.0.0",
    "fairlearn>=0.10.0",
    "imbalanced-learn>=0.12.0",
    "stable-baselines3>=2.3.0",
    "torch>=2.2.0",
    "gymnasium>=0.29.0",
    "optuna>=3.6.0",
    "tensorboard>=2.16.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    "ollama>=0.2.0",
    "prometheus-client>=0.20.0",
    "mlflow>=2.12.0",
    "python-dotenv>=1.0.0",
    "loguru>=0.7.0",
    "rich>=13.0.0",
    "pytest>=8.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["creditlens*"]
